Move Get/Set methods to appropriate file
